RUMM-2938 Expose some internal types
