+++ Datadog/Example/AppConfiguration.swift
@@ -6,7 +6,6 @@
 
 import UIKit
 import Datadog
-import DatadogInternal
 import DatadogCrashReporting
 
 protocol AppConfiguration {
